cargo clippy

+++ src/utils/bencher.rs
@@ -9,7 +9,7 @@ impl QPS for std::time::Instant {
         let time = self.elapsed();
         println!(
             "use QPS: {} QPS/s",
-            (total as u128 * 1000000000 as u128 / time.as_nanos() as u128)
+            (total as u128 * 1000000000_u128 / time.as_nanos() as u128)
         );
     }
